Code for EID Class
This is a piece of JavaScript I wrote for EID class. We made an address
book of all our customers in a Google Spreadsheet. This code, which
runs on Google servers, sends a mass email to each of our subscribers
one-by-one. This saved as a lot of time and also a lot of money if we
had used a paid subscription service like MailChimp